requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Deprecated==1.2.13
distro==1.8.0
Django==4.1.7
django-appconf==1.0.5
django-compressor==4.3.1
django-libsass==0.9
django-tinymce==3.5.0
idna==3.4
libsass==0.22.0
netaddr==0.8.0
Pillow==9.4.0
rcssmin==1.1.1
requests==2.28.2
rjsmin==1.2.1
sqlparse==0.4.3
urllib3==1.26.14
whitenoise==6.4.0
wrapt==1.15.0
yookassa==2.3.5